[fix/search-prefix/refactor2] separate classes into separate files
